Début toolbar triangle erasing TODO : changer le nombre de modes du TriEraser et continuer la toolbar

+++ Assets/Scripts/HBP/3D/Module/objects/TriEraser.cs
@@ -54,8 +54,8 @@ namespace HBP.Module3D
                 m_CurrentMode = value;
             }
         }
-        private float m_Degrees = 30f;
-        public float Degrees
+        private int m_Degrees = 30;
+        public int Degrees
         {
             get
             {
